fixing the windows compiler warnings
git-svn-id: https://svn.opendtect.org/od/trunk@38539 a985855d-d384-4b45-ae33-31a410998c3c

+++ include/AttributeEngine/attribdatacubeswriter.h
@@ -38,7 +38,7 @@ public:
 			~DataCubesWriter();
 
     void		setSelection(const TrcKeySampling&,
-				     const Interval<int>&);
+				     const Interval<float>&);
 
     od_int64		nrDone() const;
     od_int64		totalNr() const;		
@@ -62,7 +62,7 @@ private:
    TypeSet<int>			cubeindices_;
 
    TrcKeySampling		tks_;
-   Interval<int>		zrg_;
+   Interval<float>		zrg_;
 };
 
 }; //namespace
